+++ app/src/main/java/ru/practicum/android/diploma/ui/about/AboutTeamScreen.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.ui.root
+package ru.practicum.android.diploma.ui.about
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -6,6 +6,6 @@ import androidx.compose.ui.res.stringResource
 import ru.practicum.android.diploma.R
 
 @Composable
-fun FavouritesVacanciesScreen() {
-    Text(text = stringResource(R.string.favourites))
+fun AboutTeamScreen() {
+    Text(text = stringResource(R.string.team))
 }
