+++ app/src/main/java/ru/practicum/android/diploma/ui/filter/FilterSettingsScreen.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.ui.root
+package ru.practicum.android.diploma.ui.filter
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -6,6 +6,6 @@ import androidx.compose.ui.res.stringResource
 import ru.practicum.android.diploma.R
 
 @Composable
-fun SearchScreen() {
-    Text(text = stringResource(R.string.vacancies_search))
+fun FilterSettingsScreen() {
+    Text(text = stringResource(R.string.filter_settings))
 }
